requirements.txt
pytest
selemium
allure-pytest
requests
jsonpickle